Delete only existing files

+++ kopeme-core/src/test/java/de/dagere/kopeme/TestUtils.java
@@ -13,18 +13,21 @@ public class TestUtils {
    public static final String TEST_ADDITION = "testAddition";
 
    public static void deleteRecursively(final File file) {
-      if (!file.isDirectory()) {
-         if (!file.delete()) {
-            throw new RuntimeException("Could not delete " + file.getAbsolutePath());
-         }
-      } else {
-         for (final File subfile : file.listFiles()) {
-            deleteRecursively(subfile);
-         }
-         if (!file.delete()) {
-            throw new RuntimeException("Could not delete " + file.getAbsolutePath());
-         }
-      } 
+      if (file.exists()) {
+         if (!file.isDirectory()) {
+            if (!file.delete()) {
+               throw new RuntimeException("Could not delete " + file.getAbsolutePath());
+            }
+         } else {
+            for (final File subfile : file.listFiles()) {
+               deleteRecursively(subfile);
+            }
+            if (!file.delete()) {
+               throw new RuntimeException("Could not delete " + file.getAbsolutePath());
+            }
+         } 
+      }
+      
    }
 
    public static void cleanAndSetKoPeMeOutputFolder() {
